Bugfix: set last_snapshot_id in LogManager before do_snapshot when bootstrap
As we allow state machine to set snapshot index since 25c6488ab,
the term of the snapshot index may need to be got from LogManager.

Node::bootstrap will do_snapshot when last_index > 0 in Option,
so it must set term of the last_index to LogManager before do_snapshot.
Otherwise the term of last_index will be wrong.